pyproject.toml
[project]
name = "ai-agent-upgrade"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.85.0",
    "deepagents>=0.4.11",
    "langchain-daytona>=0.0.4",
    "python-dotenv>=1.2.2",
    "slack-sdk>=3.41.0",
    "langgraph>=0.2.0",
    "daytona-sdk>=0.10.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "langchain-openai>=1.1.11",
    "tavily-python>=0.7.23",
]
